pyproject.toml
[project]
name = "drone-graph"
version = "0.0.1"
description = "An execution substrate for AI swarms organized as a hivemind."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Dan Porder" }, { name = "Abhinav" }]
dependencies = [
    "anthropic>=0.40.0",
    "crawl4ai>=0.8.6",
    "neo4j>=5.20.0",
    "openai>=1.50.0",
    "pydantic>=2.8.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
    "typer>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.11.0",
    "pytest>=8.3.0",
    "ruff>=0.6.0",
    "types-tqdm>=4.66.0",
]

[project.scripts]
drone-graph = "drone_graph.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/drone_graph"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "UP", "C4", "SIM", "RUF"]

[tool.ruff.lint.per-file-ignores]
# Typer uses `typer.Option(...)` in parameter defaults by design.
"src/drone_graph/cli.py" = ["B008"]
"src/drone_graph/drones/runtime.py" = ["I001", "UP037", "RUF100"]
"src/drone_graph/terminal/shell.py" = ["N818", "SIM115", "SIM105"]

[tool.mypy]
strict = true
python_version = "3.12"
mypy_path = "src"
packages = ["drone_graph"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "integration: tests that require a live Neo4j instance",
]
